Ensure banners are always available on CI
Because they were only getting copied _the second time_ around to the
_site folder, which on CI would never come. And while looking at this
it became clear that I was using the caching wrong because it didn't
update properly for all the blog posts.

Now it will update with the content of the posts while restoring from
some fuzzy matching which will likely still create some good skipping
of unnecessary banners!