Check in the generated banners
All the banners created on GitHub were blank. I'm guessing it's
something about internet access since they're trying to load stuff off
the internet the way it is setup. I don't want to spend more time on
this right now so I'll just check them in.

In the future it will be worth changing so I run the local dev server
when generating so it can reference all the resources locally instead
of having to call out of the box.