Install SEO plugin and update banner generator
The banner generator was made very hackily and didn't properly
integrate with Jekyll, now it actually generates the banners
and make them known at creation time. And it populates the information
so it's available to other plugins.

The SEO plugin provides the tags I set before and more, so it provides
more than I used to have with less maintenance.