Remove ping because it hasn't been working
And at least Google has stopped this and relies purely on the sitemap.
They did discover that I updated yesterday and sent a crawler but no
idea how it figured it out. Then again, Feedburner is one of their
products now so maybe they got it there since the feed changed?