Add a first pass of hardcoded tests
Oh boy, I shouldn't have let Claude just go off and implement these.
I need to tell it to read the post, get the info out of the preamble,
and then create the assertions from that, and then do that per file
instead of hardcoding all of these tests.

Upside, it should be a nice little refactoring step, and I look forward
to doing the refactor/rewrite of the tests.

Will have to make an "experience report" of how the genie failed me on
this, because it was definitely a case of that I was feeling that it
was going off the deep end, but I also didn't want to bother because I
was watching analysis of Star Trek videos on Youtube so motivation was
low to correct it in the moment.

+Doing my own analysis of _how_ it did it will allow me to better
explain how to do it, and luckily this is part of my subscription
instead of a lot of wasted money. Just used up a bunch of quota.