Add word-wrapping for text and add full-width
It doesn't make sense for it to scroll, because it's often something
natural language, so scrolling doesn't make as much as sense as it
does for code.

Also added a .full-width class for cases where the code is longer and
we want to give the maximum amount of space for showing it instead,
compared to shorter snippets where the centered is nice
(which is the default).